requirements.txt
fastai
gradio
torch
pillow